% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboxcox.R
\name{myboxcox}
\alias{myboxcox}
\title{Box Cox Transformation}
\usage{
myboxcox(dat, lambda = seq(-2, 2, 1/10), iter = 1000, alpha = 0.05)
}
\arguments{
\item{dat}{a vector of data}

\item{lambda}{vector of possible values for lambda}

\item{iter}{number of iterations}

\item{alpha}{confidence level}
}
\value{
best lambda + confidence interval, plot of log likelihood
}
\description{
Box Cox Transformation
}
\examples{
myboxcox(psych$Conform)
}
