% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mychisqplot.R
\name{mychisqplot}
\alias{mychisqplot}
\title{ChiSquare Plot Test for MVN}
\usage{
mychisqplot(x)
}
\arguments{
\item{x}{a matrix of data (at least 2 variables)}
}
\value{
a chisq plot and proportion of squared distances within qchisq(0.5,p)
}
\description{
ChiSquare Plot Test for MVN
}
\examples{
mychisqplot(psych[,1:5])
}
