% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydetectout.R
\name{mydetectout}
\alias{mydetectout}
\title{Steps for Detecting Outliers}
\usage{
mydetectout(x)
}
\arguments{
\item{x}{matrix of data}
}
\value{
marginal dotplots, bivariate scatterplots, coordinates of outliers
}
\description{
Steps for Detecting Outliers
}
\examples{
mydetectout(psych[,1:5])
}
