% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myproptest.R
\name{myproptest}
\alias{myproptest}
\title{Proportion Test for Univariate Normality}
\usage{
myproptest(x)
}
\arguments{
\item{x}{a vector of univariate data}
}
\value{
proportion of data, expected proportion, and confidence intervals for 1 and 2 std deviations, and a histogram
}
\description{
Proportion Test for Univariate Normality
}
\examples{
myproptest(psych$Indep)
}
