% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myqqplot.R
\name{myqqplot}
\alias{myqqplot}
\title{My Q-Q Plot}
\usage{
myqqplot(x)
}
\arguments{
\item{x}{a vector of univariate data}
}
\value{
a Q-Q plot
}
\description{
My Q-Q Plot
}
\examples{
myqqplot(psych$Benev)
}
