% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myresult47.R
\name{myresult47}
\alias{myresult47}
\title{Result 4.7 Bivariate Normality Test}
\usage{
myresult47(x1, x2)
}
\arguments{
\item{x1}{a vector of data (variable 1)}

\item{x2}{a vector of data (variable 2)}
}
\value{
expected proportion, actual proportion, and bivariate scatter plot
}
\description{
Result 4.7 Bivariate Normality Test
}
\examples{
myresult47(psych$Indep,psych$Conform)
}
