% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrq.R
\name{myrq}
\alias{myrq}
\title{My rQ Function}
\usage{
myrq(x)
}
\arguments{
\item{x}{a vector of univariate data}
}
\value{
a Q-Q plot and the value of rQ
}
\description{
My rQ Function
}
\examples{
myrq(psych$Supp)
}
