% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych.R, R/psychdata.R
\docType{data}
\name{psych}
\alias{psych}
\title{Psychological Profile Data}
\format{
A data frame with 130 rows and 7 columns:
\describe{
  \item{Indep}{independence subscore}
  \item{Supp}{Support subscore}
  \item{Benev}{Benevolence subscore}
  \item{Conform}{Conformity subscore}
  \item{Leader}{Leadership subscore}
  \item{Gender}{1=male, 2=female}
  \item{Socio}{Socioeconomic status, 1=low, 2=medium}
}

A data frame with 130 rows and 7 columns:
\describe{
  \item{Indep}{independence subscore}
  \item{Supp}{Support subscore}
  \item{Benev}{Benevolence subscore}
  \item{Conform}{Conformity subscore}
  \item{Leader}{Leadership subscore}
  \item{Gender}{1=male, 2=female}
  \item{Socio}{Socioeconomic status, 1=low, 2=medium}
}
}
\source{
Textbook

Textbook
}
\usage{
psych

psych
}
\description{
Contains scores of teens on 5 different subsections of psychological test

Contains scores of teens on 5 different subsections of psychological test
}
\keyword{datasets}
